backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
openai
openai-whisper
requests